AgentChat-new/AzureAISearch-S3-Indexer-main/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azureaisearch-s3-indexer"
version = "0.1.0"
description = "Index S3 documents into Azure AI Search with Azure OpenAI embeddings"
authors = [{ name = "You" }]
requires-python = ">=3.9"
dependencies = [
  "boto3>=1.34.0",
  "azure-search-documents>=11.6.0b6",
  "azure-identity>=1.17.0",
  "azure-core>=1.30.0",
  "openai>=1.52.2",
  "pypdf>=4.3.1",
  "python-dotenv>=1.0.1",
  "tenacity>=8.3.0",
  "tqdm>=4.66.4",
  "requests>=2.32.3",
]

[project.scripts]
s3-indexer = "s3_indexer.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["s3_indexer*"]
